Add my submitted caesar for problem set 2
This is the naive implementation that will encrypt letters up to Y as
long as the key value is below 26. It technically passes the problem
set with a score over 70%, but is missing support for last letter of
the alphabet and larger key values by using the formula with `% 26`.

So yay 🎉 I have completed 3 of 10 assignments before 2021, but want to
go back and study the casting of characters so I can fully implement
the cipher to the problem set specification.